Rename compileRoomTypesTuple to compileCustomTypes
