+++ package.json
@@ -10,9 +10,9 @@
     "lint:ts": "eslint --color --ext .ts,.js .",
     "lint:yaml": "yamllint --ignore=node_modules/**/*.yaml **/*.yaml",
     "schemas:compile-base-types": "json2ts --cwd=src/utils/schemas/normalized/ src/utils/schemas/normalized/gameState.json > src/utils/types/gameState.d.ts",
-    "schemas:compile-custom-types": "ts-node scripts/compileRoomTypesTuple.ts",
+    "schemas:compile-custom-types": "ts-node scripts/compileCustomTypes.ts",
     "schemas:denormalize": "ts-node scripts/denormalizeSchemas.ts",
-    "schemas:prettierify": "prettier --loglevel silent --write src/utils/types/gameState.d.ts src/utils/types/roomTypesTuple.ts",
+    "schemas:prettierify": "prettier --loglevel silent --write src/utils/types/gameState.d.ts src/utils/types/customTypes.ts",
     "schemas:update": "npm run schemas:denormalize && npm run schemas:compile-base-types && npm run schemas:compile-custom-types && npm run schemas:prettierify",
     "test:custom-eslint-rules": "mocha -c --config eslint-local-rules/tests/.mocharc.js \"eslint-local-rules/tests/rules/**/*.test.ts\"",
     "test:e2e": "mocha -c \"tests/e2e/**/*.test.ts\"",
